refactor(cl-stf): use zkvm abstraction in guest
