+++ crates/proof-impl/cl-stf/Cargo.toml
@@ -7,11 +7,12 @@ version = "0.1.0"
 strata-chaintsn.workspace = true
 strata-crypto.workspace = true
 strata-primitives.workspace = true
-strata-proofimpl-checkpoint.workspace = true
 strata-proofimpl-evm-ee-stf.workspace = true
 strata-state.workspace = true
+strata-zkvm.workspace = true
 
 borsh.workspace = true
+bincode.workspace = true
 serde.workspace = true
 thiserror.workspace = true
