+++ provers/sp1/guest-cl-stf/Cargo.toml
@@ -6,16 +6,8 @@ version = "0.1.0"
 [workspace]
 
 [dependencies]
-bincode = "1.3.3"
-bitcoin = { version = "0.32.1", features = ["serde"] }
-borsh = { version = "1.5.0", features = ["derive"] }
-strata-primitives = { path = "../../../crates/primitives" }
-strata-proofimpl-checkpoint = { path = "../../../crates/proof-impl/checkpoint" }
 strata-proofimpl-cl-stf = { path = "../../../crates/proof-impl/cl-stf" }
-strata-proofimpl-evm-ee-stf = { path = "../../../crates/proof-impl/evm-ee-stf" }
-
-sha2 = "0.10.8"
-sp1-zkvm = { version = "3.0.0", features = ["verify"] }
+strata-sp1-adapter = { path = "../../../crates/zkvm/adapters/sp1" }
 
 [patch.crates-io]
 secp256k1 = { git = "https://github.com/sp1-patches/rust-secp256k1", branch = "patch-secp256k1-v0.29.1" }
